+++ pybillomat/http.py
@@ -20,7 +20,9 @@ class Connection(urllib3.HTTPSConnectionPool):
     ):
 
         # Base URL
-        url = "https://{billomat_id}.billomat.net/".format(billomat_id = billomat_id)
+        url = "https://{billomat_id}.billomat.net/".format(
+            billomat_id = billomat_id
+        )
 
         # Headers
         headers = {
@@ -32,7 +34,6 @@ class Connection(urllib3.HTTPSConnectionPool):
             headers["X-AppId"] = billomat_app_id
         if billomat_app_secret:
             headers["X-AppSecret"] = billomat_app_secret
-
 
         # Initialize ConnectionPool
         scheme, host, port = urllib3.get_host(url)